auth pages 수정
