+++ MemoWithTags/MemoWithTags/Presentation/ViewModel/SignupViewModel.swift
@@ -25,26 +25,15 @@ final class SignupViewModel: BaseViewModel, ObservableObject {
         isValidPasswordFormat = containsUppercase && containsLowercase && containsNumber && containsSpecialCharacter
     }
     
-    ///정규식으로 이메일 형식 검사
-    func checkEmailValidity(email: String) -> Bool {
-        let emailRegex = "^[A-Z0-9a-z._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,64}$"
-        let emailPredicate = NSPredicate(format: "SELF MATCHES %@", emailRegex)
-        return emailPredicate.evaluate(with: email)
-    }
-    
     func signup(nickname: String, email: String, password: String, passwordRepeat: String) async {
-        let isEmailValid = checkEmailValidity(email: email)
         checkPasswordValidity(password: password)
         let isPasswordSame = password == passwordRepeat
         
         guard !isLoading else { return }
-        
+
         if nickname.count > 8 {
             appState.system.showAlert = true
             appState.system.errorMessage = "닉네임은 8자 이하입니다."
-        } else if !isEmailValid {
-            appState.system.showAlert = true
-            appState.system.errorMessage = RegisterError.invalidEmail.localizedDescription()
         } else if !isValidPasswordFormat || !isValidLength {
             appState.system.showAlert = true
             appState.system.errorMessage = RegisterError.invalidPassword.localizedDescription()
